MIM-47 拆分 turn start 终态观察职责
